sparrow-ml/llm/requirements.txt
llama-index==0.10.12
llama-index-embeddings-langchain==0.1.2
llama-index-llms-ollama==0.1.2
llama-index-vector-stores-weaviate==0.1.3
langchain==0.1.9
sentence-transformers
weaviate-client==3.26.2
pypdf
python-box
# unstructured[all-docs]==0.12.4
typer[all]
fastapi==0.108.0
uvicorn[standard]
ollama-haystack==0.0.4
haystack-ai==2.0.0b8
ollama==0.1.6
# add pdf2htmlEX support for RAG pipeline with LlamaIndex and Unstructured

# LlamaIndex upgrade:
# pip uninstall llama-index
# pip install llama-index --upgrade --no-cache-dir --force-reinstall